feat: handle hardware deletion and optimize medication schedule logic
- Add hardware soft-delete handler to restore and update existing NFC tags

- Refactor database schema to support multiple intake times per single medication schedule

- Separate NFC tag management logic into TagMappingService

- Remove redundant if-else builder code in MedicationScheduleService

+++ Server/pill-tap-server/src/main/java/pilltapserver/domain/medicationschedule/MedicationScheduleRepository.java
@@ -2,5 +2,5 @@ package pilltapserver.domain.medicationschedule;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MedicationScheduleRepository extends JpaRepository<MedicationSchedule, Long>{
+public interface MedicationScheduleRepository extends JpaRepository<MedicationSchedule, Integer>{
 }
